pyproject.toml
[project]
name = "automation"
version = "1.0.0"
package-mode = false
requires-python = "==3.13.*"
dependencies = [
  # "easyocr==1.7.1",
  "fan-tools==3.0.0",
  "ipython",
  #"mss==10.1.0",
  "numpy==2.2.6",
  "opencv-python==4.12.0.88",
  "pyautogui==0.9.54",
  "pytest==6.2.4",
  #"system_hotkey==1.0.3",
  #"xcffib==1.5.0",
  #"xpybutil==0.0.6",
  "flake8==3.9.2",
  #"xcffib==1.5.0",
  "pillow>=12.0.0",
]

[tool.ruff]
line-length = 100
select = [
  "E",     # pycodestyle
  "F",     # pyflakes
  "I",     # isort
  "G",     # flake8-logging-format
  "ASYNC", # flake8-async
  "PIE",   # flake8-pie
  "T20",   # flake8-print
  "PT",    # flake8-pytest-style
  "Q",     # flake8-quotes
]
ignore = [
  "Q001",
  "Q003",
  "G004",
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ['snapshot_manager']
lines-after-imports = 2

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "single"

[tool.pyright]

#strict = ['capture']
